Refactor: extract data to the `contants` file
